dashboard_metrics/requirements.txt
appdirs==1.4.4
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.2
cssselect2==0.8.0
et_xmlfile==2.0.0
fonttools==4.58.0
greenlet==3.2.2
idna==3.10
importlib_metadata==8.7.0
msal==1.32.3
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
playwright==1.52.0
pycparser==2.22
pydyf==0.11.0
pyee==13.0.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
smartsheet-python-sdk==3.0.5
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
webencodings==0.5.1
websockets==10.4
zipp==3.21.0
zopfli==0.2.3.post1
